Fastdds porting - Part 1. (#39)
**Changelog:**
- basic test with fastdds + some abstraction
- Major changes to devcontainer, finally.

Merging only because of dev container


----
**Cheatsheet**
- Running status checks -> Comment PR with :pig_nose:

+++ .devcontainer/scripts/bootstrap.sh
@@ -3,22 +3,13 @@
 set -e
 
 BIN_DIR=$HOME/.local/bin
-WS_DIR=$1
+cd "$BIN_DIR"
 
-if [ ! -d "$BIN_DIR" ]; then
-    mkdir -p $BIN_DIR
+if [ ! -f "starpls" ]; then
+    wget -O starpls.tar.gz https://github.com/withered-magic/starpls/releases/download/v0.1.14/starpls-linux-amd64.tar.gz
+    tar -xf starpls.tar.gz && rm -rf starpls.tar.gz
+    chmod +x starpls
 fi
-cd $BIN_DIR
-
-# Download bazel buildifier
-wget -O buildifier https://github.com/bazelbuild/buildtools/releases/download/v7.3.1/buildifier-linux-amd64
-chmod +x buildifier
-
-# Download bazel lsp
-wget -O starpls.tar.gz https://github.com/withered-magic/starpls/releases/download/v0.1.14/starpls-linux-amd64.tar.gz
-tar -xf starpls.tar.gz && rm -rf starpls.tar.gz
-chmod +x starpls
-
 
 eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
 brew update
